backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
dj-rest-auth[with_social]==6.0.0
Django==5.1.2
django-allauth==0.61.1
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
psycopg==3.2.3
psycopg2-binary==2.9.9
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
drf-spectacular==0.27.2
pyotp==2.9.0
qrcode==8.0
Pillow==11.0.0
django-environ==0.11.2
